add charger

+++ ShamanKingProject/Assets/Scripts/MainGameMediator/PlayerStatCalculator.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 using Gamemanager;
 public static class PlayerStatCalculator 
 {
-    public static void PlayerAddOrMinusSpirit(int amount)
+    public static void PlayerAddOrMinusSpirit(float amount)
     {
         var realTimePlayerData = GameManager.Instance.MainGameMediator.RealTimePlayerData;
         realTimePlayerData.GhostSoulGageCurrentAmount = Mathf.Clamp(realTimePlayerData.GhostSoulGageCurrentAmount + amount, 0, realTimePlayerData.GhostSoulGageMaxAmount);
